repactor: Kimjunpyo's All Code Refactoring
1. 김준표가 개발한 모든 코드에 대한 리팩토링 완료
