+++ client/src/App.tsx
@@ -1,8 +1,8 @@
 import { Route, Routes, useLocation } from 'react-router-dom';
 import Main from './pages/main/Main';
-import Auth from './pages/users/Auth';
+import Auth from './pages/users/Auth/Auth';
 import Mypage from './pages/mypage/Mypage';
-import ShoppingCart from './pages/order/ShoppingCart';
+import ShoppingCart from './pages/order/ShoppingCart/ShoppingCart';
 import Store from './pages/store/Store';
 import Map from './pages/map/Map';
 import SelectStore from './pages/store/SelectStore';
