+++ client/src/assets/interface/Button.interface.ts
@@ -1,0 +1,8 @@
+import { ComponentType, ReactElement } from 'react';
+
+export interface ButtonProps {
+  title: string;
+  types: string;
+  icon?: ReactElement<ComponentType>;
+  enabled?: boolean;
+}
