+++ client/src/assets/interface/Input.interface.ts
@@ -1,0 +1,10 @@
+import { ComponentType, ReactElement } from 'react';
+
+export interface InputProps {
+  setState: React.Dispatch<React.SetStateAction<string>>;
+  state: string;
+  placeholderText: string;
+  dataValid: boolean;
+  icon: ReactElement<ComponentType>;
+  vaildMessage: string;
+}
