+++ client/src/components/CountButton.style.ts/CountButton.style.ts
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const DownBtn = styled.div`
+  cursor: pointer;
+  transition: 0.3s;
+  &:hover {
+    transform: scale(120%, 120%);
+    color: var(--white-gray);
+  }
+`;
