+++ client/src/components/Logo/Logo.style.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const LogoFont = styled.div`
+  font-family: Just Another Hand, cursive;
+  font-size: 4rem;
+  margin-bottom: 4rem;
+  cursor: pointer;
+`;
