+++ client/src/pages/users/LoginForm.tsx
@@ -1,8 +1,8 @@
 import { MdLocalPostOffice } from 'react-icons/md';
-import UserInput from '../../components/UserInput';
+import UserInput from '../../components/UserInput/UserInput';
 import { AiFillLock } from 'react-icons/ai';
 import { useEffect, useState } from 'react';
-import RoundButton from '../../components/RoundButton';
+import RoundButton from '../../components/RoundButton/RoundButton';
 import { AUTH_FAILED_MESSAGE, REGEX } from '../../assets/constantValue/constantValue';
 import { useDispatch } from 'react-redux';
 import { setAccessToken } from '../../redux/reducer/loginReducer';
